+++ src/protocol/mcpe/crypto/mod.rs
@@ -1,2 +1,2 @@
-pub mod ecdsa;
+pub mod es384;
 pub mod errors;
